Paper
We are currently in the process of submitting our paper to the International Conference on Computer Science and Computational Intelligence (ICCSCI) 2024. This conference is an internationally recognized platform for presenting the latest research findings in the field of computer science and computational intelligence. We hope that our research will make a significant contribution to the academic discussion and receive valuable feedback from experts in the field.